Address PR #94 review: collapse core/auth, drop Quart shims, simplify settings
- Empty core/auth/__init__.py; remove core/auth/cookies.py and core/auth/inputs.py
- Move cookie code into web/auth/cookies.py (Litestar-typed, no docstring, no
  Quart wrappers); drop web/auth/inputs.py — no neutral AuthInputs indirection
- web/auth/middleware.py: drop all Quart helpers (login_required,
  app_auth_required, get_current_user_from_request, _try_refresh, etc.).
  get_current_user now takes ASGIConnection directly; rename
  _try_refresh_for_connection -> public try_refresh_tokens
- web/middleware/subdomain_proxy.py: build Request/WebSocket from scope and
  reuse the connection-based helpers; drop _auth_inputs_from_scope
- web/app.py: delete _ROUTE_NAME_TO_PATH and the _url_for shim plus the
  pre-populated url_for template global; templates for unmigrated routes
  will fail at render time as intended
- Add provide_config Litestar dep; settings handlers take config: Config
- Settings routes: drop trivial attrs response classes (GetRemoteResponse,
  HostPrep, etc.); flatten host_prep fields onto top-level JSON to match
  what settings.html and test_settings_host_prep actually consume

mypy now fails on unmigrated route files that imported the removed
Quart helpers; that's intentional and they will be fixed as each route
module is ported in subsequent PRs.
